cpuidle: governors: qcom-lpm: Correctly log residency in trace
Trace ends up using int type to log which may result in negative residency
prints for large positive values. Fix by using correct type.

While at this also correct the order in TP_ARGS().

+++ drivers/cpuidle/governors/qcom-lpm.c
@@ -173,7 +173,7 @@ static void biastimer_start(uint32_t time_ns)
  * @cpu_gov:  targeted cpu's lpm data structure
  * @duration_ns:  cpu's scheduler sleep length
  */
-static uint64_t find_deviation(struct lpm_cpu *cpu_gov, int *samples_history,
+static uint64_t find_deviation(struct lpm_cpu *cpu_gov, u64 *samples_history,
 			       u64 duration_ns)
 {
 	uint64_t max, avg, stddev;
